Move section fragment out and simplify
